requirements.txt
# rot-cinn
# ========

# Requirements for the "rot-cinn" conda environment
(conda will also install all dependencies)


python 3.10.10          (includes pip but conda is recommended)
jupyterlab 3.5.3        (includes Jupyter Notebook)

freia 0.2               (via pip)

numpy 1.23.5			(in scikit-image)
scikit-learn 1.2.2      (includes scipy)
scikit-image 0.19.3
pillow 9.4.0            (in scikit-image)

pytorch 2.0.0           (-c pytorch)
torchvision 0.15.0      (-c pytorch)
If on CPU: 
	cpuonly 2.0.0       (-c pytorch)
If on Nvidia GPU: 
	pytorch-cuda 11.8   (-c nvidia)

seaborn 0.12.2          (includes pandas)
matplotlib 3.7.1		(in seaborn)
visdom 0.2.4            (-c conda-forge)
tqdm 4.65.0



# Experimental (may be temporary)
lightning-trainable 0.2.0  (via https://github.com/LarsKue/lightning-trainable)
